+++ client/index.css
@@ -41,6 +41,10 @@ header {
   border-radius: 10px;
 }
 
+.section-container {
+  width: 100%;
+}
+
 nav ul {
   list-style: none;
   display: flex;
@@ -146,6 +150,8 @@ main {
 
 .content {
   display: none;
+  width: 100%;
+
 }
 
 .content.active {
@@ -184,7 +190,7 @@ aside {
 .content {
   display: none;
   transition: all 0.3s ease-in-out;
-  
+
 }
 
 .toggle-btn {
@@ -360,7 +366,7 @@ ul.name-list li {
   border-collapse: collapse;
   padding: 10px;
   gap: 10px;
-  width: 206vh;
+  width: 100%;
   border-radius: 10px;
 }
