+++ client/index.html
@@ -11,38 +11,34 @@
   <link rel="icon" href="../client/assets/img/favicon.png" style="border-radius: 10px" type="image/x-icon" />
 </head>
 
-  <body>
-    <header>
-      <div class="logo">
-        <a href=""><img src="./assets/img/logo.png" alt="Logo" /></a>
-      </div>
-      <input type="checkbox" id="nav_check" hidden />
-      <nav>
-        <ul class="nav_menu">
-          <li><a class="link" href="#">Inicio</a></li>
-          <li>
-            <a class="link" href="#" data-target="confederaciones"
-              >Confederaciones</a
-            >
-          </li>
-          <li>
-            <a class="link" href="#" data-target="sorteoFaseGrupos"
-              >Sorteo Grupos</a
-            >
-          </li>
-          <li>
-            <a class="link" href="#" data-target="faseGrupos">Fase Grupos</a>
-          </li>
-          <li><a href="./faseFinal/faseFinal.html" target="">Fase Final</a></li>
-          <li><a href="./historial/historial.html" target="">Historial</a></li>
-        </ul>
-      </nav>
-      <label for="nav_check" class="hamburger">
-        <div></div>
-        <div></div>
-        <div></div>
-      </label>
-    </header>
+<body>
+  <header>
+    <div class="logo">
+      <a href=""><img src="./assets/img/logo.png" alt="Logo" /></a>
+    </div>
+    <input type="checkbox" id="nav_check" hidden />
+    <nav>
+      <ul class="nav_menu">
+        <li><a class="link" href="#">Inicio</a></li>
+        <li>
+          <a class="link" href="#" data-target="confederaciones">Confederaciones</a>
+        </li>
+        <li>
+          <a class="link" href="#" data-target="sorteoFaseGrupos">Sorteo Grupos</a>
+        </li>
+        <li>
+          <a class="link" href="#" data-target="faseGrupos">Fase Grupos</a>
+        </li>
+        <li><a href="./faseFinal/faseFinal.html" target="">Fase Final</a></li>
+        <li><a href="./historial/historial.html" target="">Historial</a></li>
+      </ul>
+    </nav>
+    <label for="nav_check" class="hamburger">
+      <div></div>
+      <div></div>
+      <div></div>
+    </label>
+  </header>
 
   <main>
     <div class="section-container">
@@ -88,60 +84,54 @@
     </div>
   </main>
 
-    <footer>
-      <div class="footer-content">
-        <ul class="github-list">
-          <li>
-            <a href="https://github.com/Leonardo-Galindez" target="_blank"><i class="fab fa-github"></i> Leonardo
-          </li>
-          <li>
-            <a href="https://github.com/perfil2" target="_blank"
-              ><i class="fab fa-github"></i> Perfil GitHub Juan Mamani</a
-            >
-          </li>
-          <li>
-            <a href="https://github.com/kevinquintero11" target="_blank"
-              ><i class="fab fa-github"></i> Kevin Quintero</a>
-          </li>
-          <li>
-            <a href="https://github.com/mcjaben" target="_blank"
-              ><i class="fab fa-github"></i> Benjamín Morales</a>
-          </li>
-        </ul>
-        <ul class="linkedin-list">
-          <li>
-            <a href="https://www.linkedin.com/in/leonardo-galindez-942b49291/" target="_blank"><i
+  <footer>
+    <div class="footer-content">
+      <ul class="github-list">
+        <li>
+          <a href="https://github.com/Leonardo-Galindez" target="_blank"><i class="fab fa-github"></i> Leonardo
+        </li>
+        <li>
+          <a href="https://github.com/perfil2" target="_blank"><i class="fab fa-github"></i> Perfil GitHub Juan
+            Mamani</a>
+        </li>
+        <li>
+          <a href="https://github.com/kevinquintero11" target="_blank"><i class="fab fa-github"></i> Kevin Quintero</a>
+        </li>
+        <li>
+          <a href="https://github.com/mcjaben" target="_blank"><i class="fab fa-github"></i> Benjamín Morales</a>
+        </li>
+      </ul>
+      <ul class="linkedin-list">
+        <li>
+          <a href="https://www.linkedin.com/in/leonardo-galindez-942b49291/" target="_blank"><i
               class="fab fa-linkedin"></i> Leonardo Galindez</a>
-          </li>
-          <li>
-            <a href="https://www.linkedin.com/in/perfil2" target="_blank"
-              ><i class="fab fa-linkedin"></i> Perfil LinkedIn Juan Mamani</a
-            >
-          </li>
-          <li>
-            <a
-              href="https://www.linkedin.com/in/kevin-quintero-9bb101332/"
-              target="_blank"
-              ><i class="fab fa-linkedin"></i> Kevin Quintero</a
-            >
-          </li>
-          <li>
-            <a href="https://www.linkedin.com/in/bmorales99/" target="_blank"
-              ><i class="fab fa-linkedin"></i> Benjamín Morales</a
-            >
-          </li>
-        </ul>
-        <ul class="name-list">
-          <li><i class="fas fa-user"></i> Leonardo Galindez FAI-3862</li>
-          <li><i class="fas fa-user"></i> Juan Mamani FAI-</li>
-          <li><i class="fas fa-user"></i> Kevin Quintero FAI-2771</li>
-          <li><i class="fas fa-user"></i> Benjamin Morales FAI-3370</li>
-        </ul>
-      </div>
-      <div class="footer-bottom">
-        <p>&copy; 2024 Tu Empresa. Todos los derechos reservados.</p>
-      </div>
-    </footer>
-    <script type="module" src="./index.js"></script>
-  </body>
+        </li>
+        <li>
+          <a href="https://www.linkedin.com/in/perfil2" target="_blank"><i class="fab fa-linkedin"></i> Perfil LinkedIn
+            Juan Mamani</a>
+        </li>
+        <li>
+          <a href="https://www.linkedin.com/in/kevin-quintero-9bb101332/" target="_blank"><i
+              class="fab fa-linkedin"></i> Kevin Quintero</a>
+        </li>
+        <li>
+          <a href="https://www.linkedin.com/in/bmorales99/" target="_blank"><i class="fab fa-linkedin"></i> Benjamín
+            Morales</a>
+        </li>
+      </ul>
+      <ul class="name-list">
+        <li><i class="fas fa-user"></i> Leonardo Galindez FAI-3862</li>
+        <li><i class="fas fa-user"></i> Juan Mamani FAI-</li>
+        <li><i class="fas fa-user"></i> Kevin Quintero FAI-2771</li>
+        <li><i class="fas fa-user"></i> Benjamin Morales FAI-3370</li>
+      </ul>
+    </div>
+    <div class="footer-bottom">
+      <p>&copy; 2024 Tu Empresa. Todos los derechos reservados.</p>
+    </div>
+  </footer>
+  <script src="./src/components/header.js"></script>
+  <script type="module" src="./src/index.js"></script>
+</body>
+
 </html>
